api-examples/pyproject.toml
[project]
name = "api-examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.72.0",
    "fastmcp>=2.13.0.2",
    "google-genai>=1.49.0",
    "openai>=2.7.1",
]
